refactor: move router logic into library
