+++ tests/test_ai_router.py
@@ -4,8 +4,8 @@ import contextlib
 import sys
 import pytest
 
-from scripts import ai_router
-from llm import ai_router as llm_router
+from scripts import ai_router as cli_ai_router
+from llm import router, ai_router as llm_router
 
 
 def _set_env(monkeypatch, primary="gemini", fallback="ollama"):
@@ -25,17 +25,17 @@ def test_send_prompt_uses_local_for_simple_prompt(monkeypatch):
     def mock_run_ollama(prompt, model):
         return f"ollama:{prompt}:{model}"
 
-    monkeypatch.setattr(ai_router, "run_gemini", fail_run_gemini)
-    monkeypatch.setattr(ai_router, "run_ollama", mock_run_ollama)
+    monkeypatch.setattr(router, "run_gemini", fail_run_gemini)
+    monkeypatch.setattr(router, "run_ollama", mock_run_ollama)
 
-    out = ai_router.send_prompt("hello", model="g1")
+    out = router.send_prompt("hello", model="g1")
     assert out == "ollama:hello:g1"
 
 
 def test_send_prompt_uses_primary_for_complex_prompt(monkeypatch):
     _set_env(monkeypatch, "gemini", "ollama")
 
-    long_prompt = " ".join(["word"] * (ai_router.DEFAULT_COMPLEXITY_THRESHOLD + 1))
+    long_prompt = " ".join(["word"] * (router.DEFAULT_COMPLEXITY_THRESHOLD + 1))
 
     def mock_run_gemini(prompt, model=None):
         return f"gemini:{prompt}:{model}"
@@ -43,10 +43,10 @@ def test_send_prompt_uses_primary_for_complex_prompt(monkeypatch):
     def fail_run_ollama(prompt, model):
         raise AssertionError("ollama should not be called")
 
-    monkeypatch.setattr(ai_router, "run_gemini", mock_run_gemini)
-    monkeypatch.setattr(ai_router, "run_ollama", fail_run_ollama)
+    monkeypatch.setattr(router, "run_gemini", mock_run_gemini)
+    monkeypatch.setattr(router, "run_ollama", fail_run_ollama)
 
-    out = ai_router.send_prompt(long_prompt, model="g1")
+    out = router.send_prompt(long_prompt, model="g1")
     assert out.startswith("gemini:")
 
 
@@ -59,10 +59,10 @@ def test_send_prompt_local(monkeypatch):
     def mock_run_ollama(prompt, model):
         return f"ollama:{prompt}:{model}"
 
-    monkeypatch.setattr(ai_router, "run_gemini", fail_run_gemini)
-    monkeypatch.setattr(ai_router, "run_ollama", mock_run_ollama)
+    monkeypatch.setattr(router, "run_gemini", fail_run_gemini)
+    monkeypatch.setattr(router, "run_ollama", mock_run_ollama)
 
-    out = ai_router.send_prompt("yo", local=True, model="o2")
+    out = router.send_prompt("yo", local=True, model="o2")
     assert out == "ollama:yo:o2"
 
 
@@ -76,10 +76,10 @@ def test_env_forces_remote(monkeypatch):
     def fail_run_ollama(prompt, model):
         raise AssertionError("ollama should not be called")
 
-    monkeypatch.setattr(ai_router, "run_gemini", mock_run_gemini)
-    monkeypatch.setattr(ai_router, "run_ollama", fail_run_ollama)
+    monkeypatch.setattr(router, "run_gemini", mock_run_gemini)
+    monkeypatch.setattr(router, "run_ollama", fail_run_ollama)
 
-    out = ai_router.send_prompt("short", model="g1")
+    out = router.send_prompt("short", model="g1")
     assert out == "gemini:short:g1"
 
 
@@ -93,41 +93,41 @@ def test_env_complexity_threshold(monkeypatch):
     def fail_run_ollama(prompt, model):
         raise AssertionError("ollama should not be called")
 
-    monkeypatch.setattr(ai_router, "run_gemini", mock_run_gemini)
-    monkeypatch.setattr(ai_router, "run_ollama", fail_run_ollama)
+    monkeypatch.setattr(router, "run_gemini", mock_run_gemini)
+    monkeypatch.setattr(router, "run_ollama", fail_run_ollama)
 
-    out = ai_router.send_prompt("two words", model="g1")
+    out = router.send_prompt("two words", model="g1")
     assert out == "gemini:two words:g1"
 
 
 def test_cli_invokes_send_prompt(monkeypatch):
-    def mock_send_prompt(prompt, *, local=False, model=ai_router.DEFAULT_MODEL):
+    def mock_send_prompt(prompt, *, local=False, model=router.DEFAULT_MODEL):
         assert prompt == "cli"
         assert local is True
         assert model == "m"
         return "ok"
 
-    monkeypatch.setattr(ai_router, "send_prompt", mock_send_prompt)
+    monkeypatch.setattr(router, "send_prompt", mock_send_prompt)
     out = io.StringIO()
     with contextlib.redirect_stdout(out):
-        rc = ai_router.main(["--local", "--model", "m", "cli"])
+        rc = cli_ai_router.main(["--local", "--model", "m", "cli"])
     assert rc == 0
     assert out.getvalue().strip() == "ok"
 
 
 def test_cli_reads_stdin(monkeypatch):
-    def mock_send_prompt(prompt, *, local=False, model=ai_router.DEFAULT_MODEL):
+    def mock_send_prompt(prompt, *, local=False, model=router.DEFAULT_MODEL):
         assert prompt == "from-stdin"
         assert local is False
-        assert model == ai_router.DEFAULT_MODEL
+        assert model == router.DEFAULT_MODEL
         return "done"
 
-    monkeypatch.setattr(ai_router, "send_prompt", mock_send_prompt)
+    monkeypatch.setattr(router, "send_prompt", mock_send_prompt)
     out = io.StringIO()
     stdin = io.StringIO("from-stdin")
     monkeypatch.setattr(sys, "stdin", stdin)
     with contextlib.redirect_stdout(out):
-        rc = ai_router.main(["-"])
+        rc = cli_ai_router.main(["-"])
     assert rc == 0
     assert out.getvalue() == "done\n"
 
@@ -163,10 +163,10 @@ def test_run_gemini_uses_dspy_backend(monkeypatch):
         def __init__(self, *a, **k):
             raise AssertionError("GeminiBackend should not be used")
 
-    monkeypatch.setattr(ai_router, "GeminiDSPyBackend", Dummy)
-    monkeypatch.setattr(ai_router, "GeminiBackend", Fail)
+    monkeypatch.setattr(router, "GeminiDSPyBackend", Dummy)
+    monkeypatch.setattr(router, "GeminiBackend", Fail)
 
-    out = ai_router.run_gemini("hi", model="m")
+    out = router.run_gemini("hi", model="m")
     assert out == "dspy"
     assert calls == [("init", "m"), ("run", "hi")]
 
@@ -190,10 +190,10 @@ def test_send_prompt_prefers_dspy(monkeypatch):
     def fail_ollama(prompt: str, model: str):  # pragma: no cover - ensure unused
         raise AssertionError("ollama should not be called")
 
-    monkeypatch.setattr(ai_router, "GeminiDSPyBackend", Dummy)
-    monkeypatch.setattr(ai_router, "GeminiBackend", FailBackend)
-    monkeypatch.setattr(ai_router, "run_ollama", fail_ollama)
+    monkeypatch.setattr(router, "GeminiDSPyBackend", Dummy)
+    monkeypatch.setattr(router, "GeminiBackend", FailBackend)
+    monkeypatch.setattr(router, "run_ollama", fail_ollama)
 
-    out = ai_router.send_prompt("msg", model="m")
+    out = router.send_prompt("msg", model="m")
     assert out == "dspy:msg:m"
